+++ examples/react-nextjs/app/page.tsx
@@ -2,8 +2,14 @@
 
 import dynamic from "next/dynamic";
 
-// Evolu is currently client-only. Server-side rendering support is planned for
-// the future.
+/**
+ * Evolu supports server-side rendering (SSR) even with web-only deps, but
+ * that's only practical if we want to render empty rows and don't care about
+ * the layout shift when the client hydrates.
+ *
+ * Evolu can also SSR data with server deps (check tests). Use it only for
+ * non-private (public or shared) data.
+ */
 const EvoluMinimalExample = dynamic(
   () =>
     import("@/components/EvoluMinimalExample").then(