+++ packages/react/src/index.ts
@@ -8,4 +8,4 @@ export * from "./useQueries.js";
 export * from "./useQuery.js";
 export * from "./useQuerySubscription.js";
 // export * from "./useSyncState.js"; TODO: Update it for the owner-api
-export * from "./useWasSsr.js";
+export * from "./useIsSsr.js";